Moved watchtower rpc server to rpcservers folder
