+++ rpcservers/watchtower/config_default.go
@@ -1,7 +1,7 @@
 //go:build !watchtowerrpc
 // +build !watchtowerrpc
 
-package watchtowerrpc
+package watchtower
 
 // Config is empty for non-watchtowerrpc builds.
 type Config struct{}
