+++ subrpcserver_config.go
@@ -13,7 +13,6 @@ import (
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
 	"github.com/lightningnetwork/lnd/lncfg"
-	"github.com/lightningnetwork/lnd/lnrpc/watchtowerrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/wtclientrpc"
 	"github.com/lightningnetwork/lnd/lnwire"
 	"github.com/lightningnetwork/lnd/macaroons"
@@ -28,6 +27,7 @@ import (
 	routerrpc "github.com/lightningnetwork/lnd/rpcservers/router"
 	signrpc "github.com/lightningnetwork/lnd/rpcservers/sign"
 	walletrpc "github.com/lightningnetwork/lnd/rpcservers/wallet"
+	watchtowerrpc "github.com/lightningnetwork/lnd/rpcservers/watchtower"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/watchtower"
 	"github.com/lightningnetwork/lnd/watchtower/wtclient"
